// apps/ui/airm/types/enums/node-workloads-table-field.ts
// Copyright © Advanced Micro Devices, Inc., or its affiliates.
//
// SPDX-License-Identifier: MIT

export enum NodeWorkloadsTableField {
  NAME = 'displayName',
  TYPE = 'type',
  STATUS = 'status',
  GPU_DEVICES = 'gpuDevices',
  VRAM = 'vram',
  CREATED_AT = 'createdAt',
  PROJECT = 'projectId',
}